+++ src/users/users.service.ts
@@ -3,6 +3,7 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { HashingServiceProtocol } from 'src/auth/hash/hashing.service';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class UsersService {
@@ -57,7 +58,8 @@ export class UsersService {
     }
 
 
-    async update(id: number, updateUserDto: UpdateUserDto) {
+    async update(id: number, updateUserDto: UpdateUserDto, tokenPayload: PayloadTokenDto) {
+
         try {
             const user = await this.prisma.user.findFirst({
                 where: {
@@ -67,6 +69,10 @@ export class UsersService {
 
             if (!user) {
                 throw new HttpException('Usuário não existe!', HttpStatus.BAD_REQUEST)
+            }
+
+            if (user.id !== tokenPayload.sub) {
+                throw new HttpException('Acesso negado.', HttpStatus.BAD_REQUEST)
             }
 
             const dataUser: { name?: string, passwordHash?: string } = {
@@ -96,13 +102,12 @@ export class UsersService {
             return updateUser;
 
         } catch (err) {
-            console.log(err);
             throw new HttpException('Falha ao atualizar usuário!', HttpStatus.BAD_REQUEST)
         }
     }
 
 
-    async delete(id: number) {
+    async delete(id: number, tokenPayload: PayloadTokenDto) {
         try {
             const user = await this.prisma.user.findFirst({
                 where: {
@@ -112,6 +117,10 @@ export class UsersService {
 
             if (!user) {
                 throw new HttpException('Usuário não existe!', HttpStatus.BAD_REQUEST)
+            }
+
+            if (user.id !== tokenPayload.sub) {
+                throw new HttpException('Acesso negado.', HttpStatus.BAD_REQUEST)
             }
 
             await this.prisma.user.delete({
@@ -125,7 +134,7 @@ export class UsersService {
             }
 
         } catch (err) {
-            console.log(err);
+            console.log(err)
             throw new HttpException('Falha ao deletar usuário!', HttpStatus.BAD_REQUEST)
         }
     }
